fix: change from 32 to 64

+++ internal/storage/database/documents.sql.go
@@ -41,7 +41,7 @@ type CreateDocumentParams struct {
 	Slug        string
 	Filename    string
 	MimeType    string
-	SizeBytes   int32
+	SizeBytes   int64
 	StoragePath string
 	Tags        []string
 }
